fix: code review findings — l10n, Equatable, güvenlik ve platform düzeltmeleri
- l10n: l10n.yaml oluşturuldu, eksik ARB anahtarları eklendi, AppLocalizations.delegate
  app.dart'a eklendi; tüm hardcoded string'ler context.l10n ile değiştirildi
- Equatable: Asset ve WhatIfResult domain entity'lerine Equatable eklendi
- Tip güvenliği: WhatIfResult ve model'deki num → double; fromJson'da .toDouble()
- Değişmezlik: WhatIfState listleri List.unmodifiable() ile sarıldı, const kaldırıldı
- Null güvenliği: response.data! force-unwrap → null kontrolü + FormatException
- AssetModel.fromJson: eksik/yanlış alan için açıklayıcı FormatException mesajı
- DeviceIdInterceptor: storage hatası durumunda ephemeral UUID fallback eklendi
- ApiClient: FlutterSecureStorage enjekte edilebilir parametre olarak açıldı
- Android: gradle.properties -Xmx8G → -Xmx2g, MaxMetaspaceSize kaldırıldı
- Android: .gitignore'dan gradle-wrapper.jar, gradlew, gradlew.bat kaldırıldı;
  wrapper dosyaları takip altına alındı (git add -f)
- iOS: .gitignore'dan **/dgph kaldırıldı
- iOS: RunnerTests.swift boş testExample() → XCTSkip() ile işaretlendi
- iOS: LaunchImage README.md sonuna eksik newline eklendi
- pubspec.yaml: sdk kısıtı ^3.11.1 → ^3.11.0

+++ lib/features/what_if/data/models/what_if_response_model.dart
@@ -24,13 +24,13 @@ class WhatIfResponseModel extends WhatIfResult {
         sellDate: json['sellDate'] != null
             ? DateTime.parse(json['sellDate'] as String)
             : null,
-        buyPrice: json['buyPrice'] as num,
-        sellPrice: json['sellPrice'] as num,
-        unitsAcquired: json['unitsAcquired'] as num,
-        initialValueTry: json['initialValueTry'] as num,
-        finalValueTry: json['finalValueTry'] as num,
-        profitLossTry: json['profitLossTry'] as num,
-        profitLossPercent: json['profitLossPercent'] as num,
+        buyPrice: (json['buyPrice'] as num).toDouble(),
+        sellPrice: (json['sellPrice'] as num).toDouble(),
+        unitsAcquired: (json['unitsAcquired'] as num).toDouble(),
+        initialValueTry: (json['initialValueTry'] as num).toDouble(),
+        finalValueTry: (json['finalValueTry'] as num).toDouble(),
+        profitLossTry: (json['profitLossTry'] as num).toDouble(),
+        profitLossPercent: (json['profitLossPercent'] as num).toDouble(),
         isProfit: json['isProfit'] as bool,
       );
 }
